改善 tools 可靠性
* 修正 FileSystemObject 透過 comtypes 無法使用問題
* 修正 win_exec 管理員模式 stderr 捕捉問題
* 改良 helper 例外轉送方式
* 註記 skcom 解除註冊時仍殘留登錄資訊問題
* 改善 skcom.tools.setup 例外處理方式
* 改善 skcom.tools.clean 例外處理方式
* 安裝相依套件過程寫入 setup.log
* 新增例外類別 InstallationException

Mentions: #69, #71

+++ skcom/helper.py
@@ -1,5 +1,5 @@
 """
-quicksk.helper
+skcom.helper
 """
 
 import logging
@@ -15,12 +15,14 @@ import zipfile
 
 import comtypes.client
 from comtypes import COMError
+import win32com.client
 import requests
 from packaging import version
 from requests.exceptions import ConnectionError as RequestsConnectionError
 
 from skcom.exception import ShellException
 from skcom.exception import NetworkException
+from skcom.exception import InstallationException
 
 def win_exec(cmd, admin_priv=False):
     """
@@ -38,9 +40,11 @@ def win_exec(cmd, admin_priv=False):
         while existed:
             fsn = random.randint(0, 65535)
             stdout_path = os.path.expanduser(r'~\stdout-%04x.txt' % fsn)
-            existed = os.path.isfile(stdout_path)
+            stderr_path = os.path.expanduser(r'~\stderr-%04x.txt' % fsn)
+            existed = os.path.isfile(stdout_path) or os.path.isfile(stderr_path)
         try:
             open(stdout_path, 'w').close()
+            open(stderr_path, 'w').close()
         except IOError:
             raise ShellException(-1, '無法產生 stdout 暫存檔')
 
@@ -52,6 +56,7 @@ def win_exec(cmd, admin_priv=False):
             '-FilePath', cmd[0],
             '-ArgumentList', deep_args,
             '-RedirectStandardOutput', stdout_path,
+            '-RedirectStandardError', stderr_path,
             '-NoNewWindow'
         ]
 
@@ -69,20 +74,29 @@ def win_exec(cmd, admin_priv=False):
         ]
 
         stdout = ''
+        skcom_ex = None
         try:
             # 取 stdout
-            comp = subprocess.run(cmd, check=True)
+            comp = subprocess.run(cmd, check=True, capture_output=True)
             with open(stdout_path, 'r') as stdout_file:
                 stdout = stdout_file.read()
-            # 移除暫存檔
-            os.remove(stdout_path)
         except subprocess.CalledProcessError as ex:
-            # 執行過程發生錯誤
-            raise ShellException(ex.returncode, ex.stderr.decode(charset))
+            # 執行失敗, stderr 先取表層再取底層
+            stderr = ex.stderr.decode(charset).strip()
+            if stderr == '':
+                with open(stderr_path, 'r') as stderr_file:
+                    stderr = stderr_file.read()
+            skcom_ex = ShellException(ex.returncode, stderr)
         except Exception:
             # 其他沒想到的狀況
             message = 'Unexpected exception (%s): %s' % (type(ex).__name__, str(ex))
-            raise ShellException(-1, message)
+            skcom_ex = ShellException(-1, message)
+        finally:
+            # 移除暫存檔
+            os.remove(stdout_path)
+            os.remove(stderr_path)
+            if skcom_ex is not None:
+                raise skcom_ex
     else:
         # 如果要使用 PowerShell, 含有空白字元的參數需要加上雙引號, 避免字串解析錯誤
         if cmd[0] == 'powershell.exe':
@@ -92,20 +106,24 @@ def win_exec(cmd, admin_priv=False):
 
         # 執行指令
         # 如果 shell=False, 安裝 vcredist 的時候會無法提升權限而失敗
+        skcom_ex = None
         try:
             comp = subprocess.run(cmd, check=True, shell=True, capture_output=True)
             stdout = comp.stdout.decode(charset)
         except subprocess.CalledProcessError as ex:
             # 執行過程發生錯誤
-            raise ShellException(ex.returncode, ex.stderr.decode(charset))
+            skcom_ex = ShellException(ex.returncode, ex.stderr.decode(charset))
         except FileNotFoundError as ex:
             # 沒有這個程式
             message = 'No such executable (%s)' % cmd[0]
-            raise ShellException(-1, message)
+            skcom_ex = ShellException(-1, message)
         except Exception:
             # 其他沒想到的狀況
             message = 'Unexpected exception (%s): %s' % (type(ex).__name__, str(ex))
-            raise ShellException(-1, message)
+            skcom_ex = ShellException(-1, message)
+        finally:
+            if skcom_ex is not None:
+                raise skcom_ex
 
     return stdout
 
@@ -193,6 +211,7 @@ def reg_find_value(key, value, root=winreg.HKEY_LOCAL_MACHINE):
 def verof_vcredist():
     """
     取得 Visual C++ 2010 可轉發套件版本資訊
+    注意!! 即使套件沒安裝也不可以噴例外, 否則會中斷安裝流程
     """
     try:
         node = r'SOFTWARE\WOW6432Node\Microsoft\VisualStudio\10.0\VC\VCRedist\x64:Version'
@@ -209,23 +228,32 @@ def install_vcredist():
     """
     安裝 Visual C++ 2010 x64 Redistributable 10.0.40219.325
     """
+    DEBUG_MODE = False
 
     # 下載與安裝 (安裝程式會自動切換系統管理員身分)
-    url = 'https://download.microsoft.com/download/1/6/5/' + \
-          '165255E7-1014-4D0A-B094-B6A430A6BFFC/vcredist_x64.exe'
-    vcexe = download_file(url, check_dir('~/.skcom'))
+    if not DEBUG_MODE:
+        url = 'https://download.microsoft.com/download/1/6/5/' + \
+              '165255E7-1014-4D0A-B094-B6A430A6BFFC/vcredist_x64.exe'
+        # 如果下載失敗, 會觸發 NetworkException
+        vcexe = download_file(url, check_dir('~/.skcom'))
+    else:
+        vcexe = '~/.skcom/vcredist_x64.exe'
+
+    # 如果拒絕安裝, 會觸發 ShellException "(1) 存取被拒。"
     cmd = [vcexe, 'setup', '/passive']
     win_exec(cmd)
 
-    # 等待安裝完成
-    cmd = ['tasklist', '/fi', 'imagename eq vcredist_x64.exe', '/fo', 'csv']
-    stdout = win_exec(cmd)
-    while stdout.count('\r\n') == 2:
-        time.sleep(0.5)
-        stdout = win_exec(cmd)
+    # 移除安裝包, 暫不處理檔案系統例外
+    if not DEBUG_MODE:
+        os.remove(vcexe)
 
-    # 移除安裝包
-    os.remove(vcexe)
+    # 版本檢查
+    expected_ver = version.parse('10.0.40219.325')
+    current_ver = verof_vcredist()
+    if current_ver != expected_ver:
+        raise InstallationException('Visual C++ 2010 可轉發套件', expected_ver)
+
+    return current_ver
 
 def remove_vcredist():
     """
@@ -239,7 +267,6 @@ def remove_vcredist():
     """
 
     logger = logging.getLogger('helper')
-    logger.info('移除 Visual C++ 2010 x64 Redistributable')
     cmd = [
         'msiexec.exe',
         '/X{1D8E6291-B0D5-35EC-8441-6616F567A0F7}',
@@ -250,13 +277,14 @@ def remove_vcredist():
 def verof_skcom():
     """
     檢查群益 API 元件是否已註冊
+    注意!! 即使套件沒安裝也不可以噴例外, 否則會中斷安裝流程
     """
     skcom_ver = '0.0.0.0'
     try:
         (_, dll_path) = reg_find_value(r'SOFTWARE\Classes\TypeLib', 'SKCOM.dll')
-        fso = comtypes.client.CreateObject('Scripting.FileSystemObject')
+        fso = win32com.client.Dispatch('Scripting.FileSystemObject')
         skcom_ver = fso.GetFileVersion(dll_path)
-    except COMError:
+    except Exception as ex:
         # TODO: 忘了什麼時候會炸掉
         pass
     return version.parse(skcom_ver)
@@ -268,14 +296,17 @@ def install_skcom(install_ver):
     url = 'https://www.capital.com.tw/Service2/download/api_zip/CapitalAPI_%s.zip' % install_ver
 
     # 建立元件目錄
+    # 暫不處理檔案系統例外
     com_path = check_dir(r'~\.skcom\lib')
 
     # 下載
+    # 如果下載失敗, 會觸發 NetworkException
     file_path = download_file(url, com_path)
 
     # 解壓縮
     # 只解壓縮 64-bits DLL 檔案, 其他非必要檔案不處理
     # 讀檔時需要用 CP437 編碼處理檔名, 寫檔時需要用 CP950 處理檔名
+    # 暫不處理解壓縮例外
     with zipfile.ZipFile(file_path, 'r') as archive:
         for name437 in archive.namelist():
             name950 = name437.encode('cp437').decode('cp950')
@@ -286,14 +317,35 @@ def install_skcom(install_ver):
                     extf.write(cmpf.read())
 
     # 註冊元件
+    # TODO: 使用者拒絕提供系統管理員權限時, 會觸發 ShellException,
+    #       但是登錄檔依然能寫入成功, 原因需要再調查
     cmd = ['regsvr32', r'%s\SKCOM.dll' % com_path]
     win_exec(cmd, admin_priv=True)
 
-    return True
+    # 版本檢查
+    expected_ver = version.parse(install_ver)
+    current_ver = verof_skcom()
+    if current_ver != expected_ver:
+        raise InstallationException('群益 API COM 元件', expected_ver)
+
+    return current_ver
 
 def remove_skcom():
-    """
-    TODO: 解除註冊與移除 skcom 元件
+    r"""
+    解除註冊與移除 skcom 元件
+    注意!! 解除註冊後, 這些登錄檔的值仍然會殘留:
+
+    * HKLM\SOFTWARE\Classes\TypeLib\{75AAD71C-8F4F-4F1F-9AEE-3D41A8C9BA5E}
+    * HKLM\SOFTWARE\Classes\WOW6432Node\TypeLib\{75AAD71C-8F4F-4F1F-9AEE-3D41A8C9BA5E}
+    * HKLM\SOFTWARE\WOW6432Node\Classes\TypeLib\{75AAD71C-8F4F-4F1F-9AEE-3D41A8C9BA5E}
+    * HKLM\SOFTWARE\Classes\CLSID\{54FE0E28-89B6-43A7-9F07-BE988BB40299}
+    * HKLM\SOFTWARE\Classes\CLSID\{72D98963-03E9-42AB-B997-BB2E5CCE78DD}
+    * HKLM\SOFTWARE\Classes\CLSID\{853EC706-F437-46E2-80E0-896901A5B490}
+    * HKLM\SOFTWARE\Classes\CLSID\{AC30BAB5-194A-4515-A8D3-6260749F8577}
+    * HKLM\SOFTWARE\Classes\CLSID\{E3CB8A7C-896F-4828-85FC-8975E56BA2C4}
+    * HKLM\SOFTWARE\Classes\CLSID\{E7BCB8BB-E1F0-4F6F-A944-2679195E5807}
+
+    殘留值可能會導致下次安裝時誤判已註冊
     """
     com_path = os.path.expanduser(r'~\.skcom\lib')
     com_file = com_path + r'\SKCOM.dll'
@@ -301,7 +353,6 @@ def remove_skcom():
         return
 
     logger = logging.getLogger('helper')
-    logger.info('移除群益 API 元件')
     logger.info('  路徑: %s', com_path)
     logger.info('  解除註冊: %s', com_file)
     cmd = ['regsvr32', '/u', com_file]
@@ -309,7 +360,6 @@ def remove_skcom():
 
     logger.info('  移除元件目錄')
     shutil.rmtree(com_path)
-
 
 def has_valid_mod():
     r"""
@@ -378,7 +428,6 @@ def clean_mod():
             continue
 
         logger = logging.getLogger('helper')
-        logger.info('移除 comtypes 套件自動生成檔案')
         logger.info('  路徑 %s', gendir)
 
         for item in os.listdir(gendir):
